Accelerate reconcile using Perforce digests
Instead of copying every file from the source repo to the destination
repo and then running `p4 revert -a`, switch to comparing the
Perforce-provided digest and only copying the files whose digest, type,
or case has changed.

1. Replaces `p4 -ztag files -e` with
   `p4 fstat -T depotFile,headAction,headChange,headType,digest -Ol -F '^(headAction=move/delete | headAction=purge | headAction=archive | headAction=delete)'`
    (which contains the same information, though it also includes the digest)

2. Updates `p4.runAndParseDepotFiles` to support the `head*` names
   of fields and also parse `digest`

3. Modifies `Reconcile` to compare digests and only adding files to
   `Match` if they're actually different in some way

This also has the benefit of not returning that there are no changes if
only existing file contents has changed, not paths or types.